desperately renaming util.google in case it solves crazy bug
